requirements.txt
numpy
joblib
tensorflow
pillow
streamlit